refactor(test): 테스트 코드 MVC 리펙터링
테스트 코드 MVC 구조로 리펙터링
- Ball과 Balls는 model로 이동
- InputViewTest 생성
