+++ src/test/java/baseball/model/BallTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.model;
 
 import baseball.model.Ball;
 import baseball.model.PlayType;
