+++ src/test/java/baseball/view/InputViewTest.java
@@ -1,0 +1,9 @@
+package baseball.view;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class InputViewTest {
+
+}
